--- src/pages/RestoreAccount/RestoreAccount.css
@@ -1,45 +1,0 @@
-.restore-account-form {
-  margin-top: 36px;
-}
-
-.title-restore {
-  font-size: 24px;
-  font-weight: 500;
-  margin: 2rem 0 2rem 0;
-}
-
-.restore-button-wrapper {
-  display: grid;
-  gap: 0.5rem;
-  justify-content: center;
-}
-
-.restore-button {
-  width: 300px;
-}
-
-.creating-loading-warapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 50vh;
-}
-
-.loading-big {
-  width: 190px;
-  height: 190px;
-}
-
-.loading-big::after {
-  width: 160px;
-  height: 160px;
-  border: 15px solid #fff;
-  border-color: rgb(var(--color-main-green)) transparent rgb(var(--color-black))
-    transparent;
-}
-
-.loadingText {
-  margin-bottom: 60px;
-  font-size: 24px;
-  font-weight: 400;
-}
