+++ src/pages/RestoreAccount/RestoreAccount.module.css
@@ -1,0 +1,55 @@
+.page {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px;
+}
+
+.title {
+  font-size: 24px;
+  font-weight: 700;
+  color: rgb(var(--color-black));
+  margin-bottom: 8px;
+  text-align: center;
+}
+
+.subtitle {
+  font-size: 14px;
+  color: rgb(var(--color-dark-gray));
+  text-align: center;
+  margin-bottom: 28px;
+}
+
+.cards {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  width: 100%;
+  max-width: 400px;
+}
+
+.loadingWrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 50vh;
+}
+
+.loadingBig {
+  width: 190px;
+  height: 190px;
+}
+
+.loadingBig::after {
+  width: 160px;
+  height: 160px;
+  border: 15px solid #fff;
+  border-color: rgb(var(--color-main-green)) transparent rgb(var(--color-black))
+    transparent;
+}
+
+.loadingText {
+  margin-bottom: 60px;
+  font-size: 24px;
+  font-weight: 400;
+}
